+++ test/dummy/db/seeds.rb
@@ -79,7 +79,7 @@ begin
       parent_recording: terms_recording,
       attributes: { slug: sample_slug, status: "published" }
     ).value!
-  elsif terms_recording.recordable.title != sample_title
+  elsif terms_recording.recordable.title != sample_title || terms_recording.recordable.body != sample_body
     terms_recording = root_recording.revise(terms_recording, actor: user) do |terms|
       terms.title = sample_title
       terms.body = sample_body