+++ test/dummy/test/integration/accept_terms_test.rb
@@ -31,7 +31,7 @@ class AcceptTermsTest < ActionDispatch::IntegrationTest
     assert_includes response.body, "Studio Terms"
     assert_includes response.body, "Be kind"
     assert_includes response.body, "I agree to these terms"
-    assert_includes response.body, "flat-pack-content-editor-content"
+    assert_includes response.body, "fp-content"
     published_on = @recording.current_publishable.publish_at.in_time_zone.strftime("%e %b %Y").squish
     assert_includes response.body, published_on
     refute_includes response.body, "ago"